styling: about me homepage basic layout
- Separate components for editing/displaying about-me info
