+++ client/styles/homepage-aboutme.css
@@ -1,0 +1,151 @@
+/* about-me.css */
+.about-me {
+  padding: 2rem 5%;
+  background-color: #f8f9fa;
+  color: #333333;
+  min-height: 85vh;
+  min-width: 75vw;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+}
+
+.about-me__container {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  max-width: 1000px;
+  gap: 2rem;
+  flex-wrap: wrap;
+  width: 100%;
+}
+
+/* Profile Picture */
+.about-me__profile-picture {
+  flex: 1;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  min-width: 150px;
+}
+
+.about-me__profile-img {
+  width: 100%;
+  max-width: 150px;
+  aspect-ratio: 1/1;
+  border-radius: 50%;
+  border: 0.25rem solid #A6C8FF;
+  object-fit: cover;
+}
+
+/* Text Content */
+.about-me__content {
+  flex: 2;
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  max-width: 100%;
+}
+
+/* Header */
+.about-me__header {
+  margin-bottom: 1rem;
+}
+
+.about-me__name {
+  font-size: 2rem;
+  font-weight: bold;
+  color: #333333;
+}
+
+.about-me__role {
+  font-size: 1.5rem;
+  color: #555555;
+}
+
+.about-me__location {
+  font-size: 1rem;
+  color: #666666;
+}
+
+/* Contact Information */
+.about-me__contacts {
+  display: flex;
+  gap: 1rem;
+  flex-wrap: wrap;
+  margin-bottom: 1rem;
+}
+
+.about-me__contact-link {
+  color: #A6C8FF;
+  text-decoration: none;
+  font-weight: bold;
+}
+
+.about-me__contact-link:hover {
+  text-decoration: underline;
+}
+
+/* Favorite Technologies */
+.about-me__technologies {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.about-me__tech-icon {
+  padding: 0.3rem 0.6rem;
+  background-color: #A6C8FF;
+  color: white;
+  border-radius: 4px;
+  font-size: 0.8rem;
+}
+
+/* Scroll Indicator */
+.about-me__scroll-indicator {
+  position: absolute;
+  bottom: 10px;
+  left: 50%;
+  transform: translateX(-50%);
+  font-size: 1rem;
+  color: #555555;
+  cursor: pointer;
+  animation: bounce 2s infinite;
+}
+
+@keyframes bounce {
+  0%, 20%, 50%, 80%, 100% { transform: translateX(-50%) translateY(0); }
+  40% { transform: translateX(-50%) translateY(-10px); }
+  60% { transform: translateX(-50%) translateY(-5px); }
+}
+
+/* Responsive Design for Tablets */
+@media (max-width: 768px) {
+  .about-me__container {
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+  }
+
+  .about-me__contacts {
+    justify-content: center;
+  }
+
+  .about-me__profile-picture {
+    margin-bottom: 1rem;
+  }
+
+  .about-me__profile-img {
+    max-width: 120px;
+  }
+
+  .about-me__name {
+    font-size: 1.8rem;
+  }
+
+  .about-me__role {
+    font-size: 1.
+  }
+}
